update icu: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@44815 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/icu/samples/udata/writer.c
@@ -2,7 +2,7 @@
 *******************************************************************************
 *
 *   © 2016 and later: Unicode, Inc. and others.
-*   License & terms of use: http://www.unicode.org/copyright.html#License
+*   License & terms of use: http://www.unicode.org/copyright.html
 *
 *******************************************************************************
 *******************************************************************************
@@ -70,7 +70,7 @@ main(int argc, const char *argv[]) {
     uint16_t intValue=2000;
     
     long dataLength;
-    uint32_t size;
+    size_t size;
 #ifdef WIN32
     char *currdir = _getcwd(NULL, 0);
 #else
@@ -107,19 +107,8 @@ main(int argc, const char *argv[]) {
 
 
     if(dataLength!=(long)size) {
-        fprintf(stderr, "Error: data length %ld != calculated size %lu\n", dataLength, size);
+        fprintf(stderr, "Error: data length %ld != calculated size %zu\n", dataLength, size);
         exit(U_INTERNAL_PROGRAM_ERROR);
     }
     return 0;
 }
-
-
-
-
-
-
-
-
-
-
-
